Separate index-window state
