+++ client/components/window-contents/IndexWindowContentComponent.js
@@ -29,8 +29,8 @@ export default class IndexWindowContentComponent extends React.Component {
     return (
       <box { ...props } >
       {
-        this.props.window.listPagination.indexedObjects.map(({ serialNumber, object }, idx) => {
-          const isCursorOver = this.props.window.cursorIndex === idx;
+        this.props.indexWindow.listPagination.indexedObjects.map(({ serialNumber, object }, idx) => {
+          const isCursorOver = this.props.indexWindow.cursorIndex === idx;
           const rowProps = Object.assign({}, rowPropsTemplate, {
             key: 'object-' + idx,
             top: idx,
